refacor(nc-gui): form view code
